Refactor set-password
