+++ src/pages/authentication/index.ts
@@ -1,6 +1,6 @@
 export {default as LoginForm} from './login/login';
 export {default as RegisterForm} from './register/register';
 export {default as RequestPasswordForm} from './request-password/request-password';
-export {default as SetPasswordForm} from './SetPasswordForm';
+export {default as SetPasswordForm} from './set-password/set-password';
 export {default as VerifyForm} from './VerifyForm';
 export {default as VerifyWall} from './VerifyWall';
